Move Pretty to Util.Pretty and put it in other-modules.
